Make Controller instances singleton.
This would be very useful when a controller has to maintain state. It is
also less memory-intensive as only one instance of a controller is created,
which in turn could mean only one view, model, etc.

With a navigation-intensive app this could have a big impact, though it
may turn out to be a premature optimisation.