Re-organise files and remove requirement for server-side includes.
Instead of relying on server-side includes to stich all the files together
(which at the time was a good idea), include them as separate assets in
the HTML file. This improves error reporting at the cost of more requests.
For a private app this shouldn't be an issue and for production use a
build script can be introduced.