Add support for message localisation.
Using I18n.language(..) a desired language can be specified for all messages
that undergo translation. There must be a module defined that matches
'I18n/{language}' and it must return a hash object.

If no such module is available, a graceful fallback to the original message
will be assumed.