Instead of using plain CSS, use Less.js instead.
While this is not as good as being able to have the full power of Sass/Compass
on the client-side, it is a good substitute and should allow for less noise
in the stylesheets.

In order to expose Less.js as an AMD module, fake AMD support in the module
definition function.