Add {Route,View}.get(..) for navigating to a specific controller/action.
Instead of fixing all URLs in code, use the controller/action to infer the
URL. This allows us to rapidly change the URLs without modifying any code.

To make this work, we need to be able to convert a regular expression to a
string which we can pass on to Backbone. RandExp.js does just that, but
it won't deal with groups for us. This functionality is missing and will be
added at a later point as needed.