Add Route.previous(..) to replace history.back(..).
Route.previous(..) allows custom `params` to be attached to the request so
the previous page can restore its context properly. This would rarely be used,
but the method is nevertheless important as it ensures all navigation in the
app happens through the Route object.

The reference translation was updated as well.