Add { navigate: false } option to Route.get.
When specified, an internal request will be made and the specified
controller and action will be fired without changing the URL in the browser.
This is ideal for cases when a controller fails and must revert to the state
of a previous controller/action, such as for example a log in attempt with
no information posted.