Override Backbone's navigate(..) method to process the URL even if no change.
When we call Route.{get,post}(..) with the same URL as the current one,
we need to override Backbone's behaviour and trigger an update regardless.
Internall, Backbone tracks the fragment from the URL and will not call
loadUrl(..) unless there was a change. By telling it to never check for
changes ({ trigger: false }), we can override this default behaviour and
load the URL regardless.