Add a basic Model class.
The Model requires that all attributes are defined in the `defaults` hash.
If an attempt is made to set an attribute which is not defined inside the
hash, the key will be excluded.

Models must also contain a 'name' which is used to build the sync URL.

Update reference translation (Bulgarian) as well.