Add proper Deferred support for Model sync methods.
The existing Backbone implementation returns the underlying jQuery.ajax
result. This is a Deferred object which can be resolved/rejected. However,
when registering a .done(..) callback, for example, access to the model
will not be available as it would be if { success: ... } was used. By
overriding the Deferred object we take care of that.