Unbind events from an element if another View is being rendered.
This is needed to avoid events from Views that no longer exist to fire.
When a View is rendered, it will also register itself on the element.
If another View tried to render itself on the same element, it will first
check to see if there was a previous view. If this is the case, all events
will be deattached.